refactor(phpstan): Implement PHPStan lvl 3
