+++ src/Handler/FormErrorHandler.php
@@ -141,6 +141,7 @@ final class FormErrorHandler implements SubscribingHandlerInterface
 
     private function convertFormToArray(SerializationVisitorInterface $visitor, FormInterface $data): \ArrayObject
     {
+        /** @var \ArrayObject{errors?: string[], children?: array<string, \ArrayObject>} $form */
         $form = new \ArrayObject();
         $errors = [];
         foreach ($data->getErrors() as $error) {
