+++ tests/Metadata/ClassMetadataTest.php
@@ -32,8 +32,8 @@ class ClassMetadataTest extends TestCase
     {
         $meta = new ClassMetadata('JMS\Serializer\Tests\Metadata\PropertyMetadataOrder');
         $meta->xmlRootPrefix = 'foo';
-        $meta->xmlDiscriminatorCData = 'foo';
-        $meta->xmlDiscriminatorAttribute = 'foo';
+        $meta->xmlDiscriminatorCData = true;
+        $meta->xmlDiscriminatorAttribute = false;
         $meta->xmlRootName = 'foo';
 
         $restoredMeta = unserialize(serialize($meta));
